+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 2.1.0 (2026-04-13) :
+- added 2 commands to manage history mode :
+	- show_history
+	- clean_history
+- added tests for history mode
+- improved filesystem mocking system for testing purpose (FSMock)
+
 ## v2.0.0 (2026-04-08) :
 - added history mode :
 	- files are saved in an history mode directory before any update